вынес форму с подпиской на новости в отдельный компонент SupportForm
